+++ devices/switch/on_off_switch.py
@@ -7,10 +7,16 @@ class OnOffSwitch(Device):
         return Domoticz.Device(Unit=unit, DeviceID=device_id, Name=device_name, TypeName="Switch", Image=1).Create()
 
     def get_numeric_value(self, value, device):
-        return 1 if value.lower() == 'on' else 0
+        if (value.lower() == 'on'):
+            return 1
+        elif (value.lower() == 'off'):
+            return 0
+        else:
+            return device.nValue
 
     def get_string_value(self, value, device):
-        return 'On' if value.lower() == 'on' else 'Off'
+        n_value = self.get_numeric_value(value, device)
+        return 'On' if n_value == 1 else 'Off'
 
     def handle_command(self, device_data, command, level, color):
         device_address = device_data['ieee_addr']